рефакторинг: основные методы вынесены в статический класс
